+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `cache_control` and `acl` parameters to S3 storage for CloudFront (gshaw)
+
 * Improve Windows compatibility in the FileSystem storage (janko-m)
 
 * Remove the ability for FileSystem storage to accept IDs starting with a slash (janko-m)